requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pyjwt>=2.8.0
requests>=2.31.0